requirements.txt

Flask
Pillow
piexif
